+++ py/packages/sdk/src/antfly/client_generated/models/__init__.py
@@ -812,8 +812,6 @@ from .inference_read_response_object import InferenceReadResponseObject
 from .inference_read_result import InferenceReadResult
 from .inference_read_result_fields import InferenceReadResultFields
 from .inference_request_admission_config import InferenceRequestAdmissionConfig
-from .inference_rerank_multimodal_document import InferenceRerankMultimodalDocument
-from .inference_rerank_multimodal_request import InferenceRerankMultimodalRequest
 from .inference_rerank_object import InferenceRerankObject
 from .inference_rerank_object_object import InferenceRerankObjectObject
 from .inference_rerank_request import InferenceRerankRequest
@@ -2091,8 +2089,6 @@ __all__ = (
     "InferenceReadResult",
     "InferenceReadResultFields",
     "InferenceRequestAdmissionConfig",
-    "InferenceRerankMultimodalDocument",
-    "InferenceRerankMultimodalRequest",
     "InferenceRerankObject",
     "InferenceRerankObjectObject",
     "InferenceRerankRequest",